html: Improve display of the game-round labels (ITM17)
Improves the display of the table headers which provide labels for the
game rounds.

Using numerical characters instead of words avoids possible confusion
with the word 'tress'.

Using non-breaking spaces between the number and the card pattern makes
for a neater display on small screens.

A line break is used instead of '/' for a neater display as well.